Updated delete activity for only the owner of activity or the owner of the trip with that activity can delete the activity

+++ OutTheGC/Repositories/ActivityRepository.cs
@@ -2,6 +2,7 @@
 using OutTheGC.Models;
 using OutTheGC.Data;
 using Microsoft.EntityFrameworkCore;
+using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace OutTheGC.Repositories;
 
@@ -86,13 +87,20 @@ public class ActivityRepository : IActivityRepository
         return updatedActivity;
     }
 
-    public async Task<Activity> DeleteActivityAsync(Guid activityId)
+    public async Task<Activity> DeleteActivityAsync(Guid activityId, Guid userId)
     {
-        var activityToDelete= await dbContext.Activities.SingleOrDefaultAsync(a => a.Id == activityId);
+        var activityToDelete = await dbContext.Activities.SingleOrDefaultAsync(a => a.Id == activityId);
 
         if (activityToDelete == null)
         {
             return null;
+        }
+
+        var tripOwnerId = await dbContext.Trips.Where(t => t.Id == activityToDelete.TripId).Select(t => t.UserId).SingleOrDefaultAsync();
+
+        if (activityToDelete.UserId != userId && tripOwnerId != userId)
+        {
+            throw new Exception("403 Forbidden: User is not authorized to delete this activity");
         }
 
         dbContext.Activities.Remove(activityToDelete);
